Fixed a few issues in Fortran User adapter introduced by the name change and handle change: - On 32 bit machines, the fortran handles had only 32 bits, causing problems with the 64 bit scorep handles. - The name change made names longer, causing conflicts with 72 char limit in fixed form fortran. Thus, renamed some functions. - Added the unexpanded test file, to have the source of the expanded test file in the repo, as well.
Subversion: revision:1055

+++ src/adapters/include/SCOREP_Fortran_Wrapper.h
@@ -108,17 +108,17 @@
 /**
     Defines the type obtained from Fortran calls for region handles.
  */
-typedef int SCOREP_Fortran_RegionHandle;
+typedef int64_t SCOREP_Fortran_RegionHandle;
 
 /**
     Defines the type obtained from Fortran calls for metric handles.
  */
-typedef int SCOREP_Fortran_MetricHandle;
+typedef int64_t SCOREP_Fortran_MetricHandle;
 
 /**
     Defines the type obtained from Fortran calls for metric group handles.
  */
-typedef int SCOREP_Fortran_MetricGroup;
+typedef int64_t SCOREP_Fortran_MetricGroup;
 
 /**
     @def SCOREP_FORTRAN_INVALID_REGION
